+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### AppCenter
 
 * **[Feature]** Now supports tvOS.
+* **[Fix]** Drop and recreate the database when it is corrupted.
 
 ### AppCenterAnalytics
 
@@ -13,6 +14,14 @@
 ### AppCenterCrashes
 
 * **[Feature]** Now supports tvOS.
+
+___
+
+## Version 2.2.0 (Under development)
+
+### AppCenterData
+
+* **[Feature]** Add support for offline list of documents.
 
 ___